requirements.txt
django>=4.0
django-mptt
jsonfield
django_jinja
celery
django-compressor
django-sortedm2m==3.1.1
django-reversion
martor
ua_parser